Further simplify the main Bot class by moving most listeners to cogs
