+++ cogs/mod_commands.py
@@ -474,63 +474,5 @@ class ModCommands(Cog):
         except discord.Forbidden:
             raise CommandError(message=f'Missing permissions: `manage_nicknames`. Or insufficient permissions to change {member.display_name}\'s nickname.')
 
-    @commands.command(aliases=['delall'])
-    @is_admin()
-    async def deleteall(self, ctx: Context, *, channel: GuildChannel) -> None:
-        '''
-        Deletes all messages that will be sent in the given channel. (Admin+)
-        Arguments: channel (mention, name, or id)
-        '''
-        self.bot.increment_command_counter()
-
-        if not channel or not ctx.guild:
-            raise CommandError(message=f'Required argument missing: `GuildChannel`.')
-
-        async with self.bot.async_session() as session:
-            guild: Guild = await get_db_guild(self.bot, ctx.guild, session)
-            if guild.delete_channel_ids:
-                if channel.id in guild.delete_channel_ids:
-                    guild.delete_channel_ids = guild.delete_channel_ids.remove(channel.id)
-                    message: str = f'Messages in {channel.mention} will no longer be deleted.'
-                else:
-                    guild.delete_channel_ids = guild.delete_channel_ids + [channel.id]
-                    message = f'All future messages in {channel.mention} will be deleted.'
-            else:
-                guild.delete_channel_ids = [channel.id]
-                message = f'All future messages in {channel.mention} will be deleted.'
-            await session.commit()
-        await ctx.send(message)
-
-    
-    @commands.command(hidden=True, aliases=['hof'])
-    @is_admin()
-    async def hall_of_fame(self, ctx: Context, *, channel: GuildChannel | None, react_num: int = 10) -> None:
-        '''
-        Sets the hall of fame channel and number of reactions for this server. (Admin+)
-        Arguments: channel (mention, name, or id), react_num (int)
-        After [react_num] reactions with the :star2: emoji, messages will be shown in the hall of fame channel.
-        '''
-        self.bot.increment_command_counter()
-
-        if not ctx.guild:
-            raise CommandError(message=f'Required argument missing: `guild`.')
-        
-        if channel and react_num < 1:
-            raise CommandError(message=f'Invalid argument: `react_num` (`{react_num}`). Must be at least 1.')
-
-        async with self.bot.async_session() as session:
-            guild: Guild = await get_db_guild(self.bot, ctx.guild, session)
-
-            if not channel:
-                guild.hall_of_fame_channel_id = None
-                message: str = f'Disabled hall of fame for this server.'
-            else:
-                guild.hall_of_fame_channel_id = channel.id
-                guild.hall_of_fame_react_num = react_num
-                message = f'Set the hall of fame channel for this server to {channel.mention}. The number of :star2: reactions required has been set to `{react_num}`.'
-            await session.commit()
-
-        await ctx.send(message)
-
 async def setup(bot: Bot) -> None:
     await bot.add_cog(ModCommands(bot))
